fix: refactor Tests per PR suggestions
